+++ index.html
@@ -3,8 +3,9 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <link rel="icon" type="image/png" href="assets/favicon.png" />
     <link rel="stylesheet" href="styles/index.css" />
-    <title>Document</title>
+    <title>storo08 chase</title>
   </head>
   <body>
     <form class="form">
@@ -22,9 +23,9 @@
       <thead>
         <tr>
           <th>Идентификатор</th>
-          <th>Игрок</th>
-          <th>Рейк в месяце</th>
-          <th>Рейкбек</th>
+          <th>Логин игрока</th>
+          <th>Рейк за месяц</th>
+          <th>Chase выплата</th>
         </tr>
       </thead>
       <tbody class="table-body"></tbody>
