+++ scripts/helpers/index.js
@@ -1,0 +1,3 @@
+export { default as datesOfMonth } from './datesOfMonth.js';
+export { default as prize } from './prize.js';
+export { default as renderData } from './renderData.js';
