+++ scripts/index.js
@@ -1,4 +1,4 @@
-import steps from './steps.js';
+import { datesOfMonth, prize, renderData } from './helpers/index.js';
 
 const tableBody = document.querySelector('.table-body');
 const form = document.querySelector('.form');
@@ -7,54 +7,15 @@ const month = document.querySelector('.month');
 const requestMonth = document.querySelector('.request-month');
 const rakeTotal = document.querySelector('.rake-total');
 
-const renderTableData = (item) => {
-  const tr = document.createElement('tr');
-
-  const id = document.createElement('td');
-  id.textContent = item.player_identificator;
-  tr.append(id);
-
-  const username = document.createElement('td');
-  username.textContent = item.username;
-  tr.append(username);
-
-  const rakeCurrent = document.createElement('td');
-  rakeCurrent.textContent = item.rake;
-  tr.append(rakeCurrent);
-
-  const rakeBackCurrent = document.createElement('td');
-  rakeBackCurrent.textContent = steps(item.rake);
-  tr.append(rakeBackCurrent);
-
-  tableBody.append(tr);
-};
-
 const date = new Date();
 year.value = date.getUTCFullYear();
 month.value = date.getUTCMonth() + 1;
-
-const dateOfMonth = (year, month) => {
-  const dayList = [];
-
-  const currentDate = date.getUTCDate();
-  const currentMonth = date.getUTCMonth() + 1;
-  const daysPerMonth =
-    currentMonth == month ? currentDate : new Date(year, month, 0).getDate();
-
-  for (let i = 0; i < daysPerMonth; i++) {
-    const monthPadStart = month.toString().padStart(2, '0');
-    const dayPadStart = (i + 1).toString().padStart(2, '0');
-    dayList.push(`${year}-${monthPadStart}-${dayPadStart}`);
-  }
-
-  return dayList;
-};
 
 form.addEventListener('submit', (event) => {
   event.preventDefault();
   tableBody.innerHTML = 'Загрузка данных...';
 
-  const dateList = dateOfMonth(year.value, month.value);
+  const dateList = datesOfMonth(year.value, month.value);
 
   const requestList = dateList.map((date) => {
     const url = `https://twister-races.onrender.com/hands?date=${date}`;
@@ -97,11 +58,11 @@ form.addEventListener('submit', (event) => {
     .then((data) => {
       let rakeTotalCount = 0;
       data
-        .filter((item) => item.rake >= 1000)
+        .filter((item) => item.rake >= 1000 && item.username != 'sanchess08')
         .sort((a, b) => b.rake - a.rake)
         .map((item) => {
-          renderTableData(item);
-          rakeTotalCount += steps(item.rake);
+          tableBody.append(renderData(item));
+          rakeTotalCount += prize(item.rake);
         });
       rakeTotal.textContent = rakeTotalCount;
     });
